COMP: fix install of dcm2niix and related
This install the superbuilt executable as part of the
extension.

Also clean up the code so the DWI volume is auto-loaded
and named after the input directory (remove extra combobox).

Some extra code cleanups and diagnostic output.

TODO: check extension to confirm this works on windows
(might need to add .exe to cmdPath string).