ENH: Add DICOM plugin that allows loading images using dcm2niix from the DICOM browser
Confidence of dcm2niix plugin is set to 0.45 (or 0.3, if there are warnings), which means that still the usual scalar volume plugin will be used to load the volume by default.
However, when loading fails, the user can enable advanced option and choose the loadable generated by dcm2niix.

Also enabled OpenJPEG and JPEG-LS for building dcm2niix to enable loading of compressed images.